stable tkinter GUI
stable version of strain_server using a tkinter+matplotlib based GUI. In this version we have removed the shutdown buttom (which wasn't working properly). It works well on Mac, but matplotlib causes a non-responsive tkinter window on Windows OS. For this reason, we will now start a development branch where the display is based on Qt.